% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\alias{app_ui}
\alias{app_server}
\title{app_ui
UI for CHI-power Dashboard}
\usage{
app_ui()

app_server()
}
\value{
UI for CHI-power Dashboard

Server function for CHI-power Dashboard
}
\description{
app_ui
UI for CHI-power Dashboard

app_server
Server for CHI-power Dashboard
}
