% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somalogic.R
\name{somalogic}
\alias{somalogic}
\alias{somalogic_sidebarUI}
\alias{somalogic_cardsUI}
\alias{somalogic_server}
\title{somalogic_sidebarUI
UI element for the SomaLogic sidebar}
\usage{
somalogic_sidebarUI(id)

somalogic_cardsUI(id)

somalogic_server(id)
}
\arguments{
\item{id}{Shiny namespace ID}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
somalogic_sidebarUI
UI element for the SomaLogic sidebar

somalogic_cardsUI
UI element for the SomaLogic cards

somalogic_server
Server element for SomaLogic tab
}
